pyproject.toml
[tool.poetry]
name = "telegram-forwarder"
version = "0.1.0"
description = ""
authors = ["vi.latyshev <storis2012@yandex.ru>"]
packages = [
    { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
Pyrogram = "^1.4.14"
TgCrypto = "^1.2.3"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src.main:start"
